requirements.txt
beautifulsoup4>=4.12.0
faiss-gpu
faiss-cpu
google-generativeai>=0.3.0
groq>=0.4.0
ipython>=8.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.6
numpy>=1.24.0
ollama>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
unstructured
